4ºano/1ºsemestre/4.1 - Requisitos e Arquiteturas de software/Fase 3/local/backend/ms-project/picturas_removebg_tool_ms/app/removebg_request_message.py
from pydantic import BaseModel


from .core.messages.request_message import RequestMessage

class RemovebgParameters(BaseModel):
    inputImageURI: str
    outputImageURI: str
    
RemovebgRequestMessage = RequestMessage[RemovebgParameters]