4ºano/1ºsemestre/4.1 - Requisitos e Arquiteturas de software/Fase 3/local/requirements.txt
boto3==1.35.97
botocore==1.35.97
jmespath==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
rembg==2.0.61
